requirement.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.0
numpy==1.14.1
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
six==1.11.0
